feat: rename negative intents service
